Image_Classification_App/requirements.txt
streamlit==0.79.0
numpy==1.20.2
pandas==1.1.5
pillow==8.0.1
pytorch==1.6.0
torchvision==0.7.0
albumentations==0.5.2
opencv-python==4.0.1
opencv-contrib-python==4.0.1
